+++ lib/game/index.ts
@@ -1,0 +1,5 @@
+import { buildTestGame } from './factory'
+import { submitAction } from './actions'
+import { processStackItem } from './stack'
+
+export { buildTestGame, submitAction, processStackItem }
